requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.20
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
marshmallow>=3.20.0
pytest>=7.4.0
python-dotenv>=1.0.0
